# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gun Del Sol API",
    "description": "High-performance async API for Solana token analysis (Modular)",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "Settings & Health"
        ],
        "summary": "Get Api Settings",
        "description": "Get current API settings\n\nReturns:\n    API settings dictionary",
        "operationId": "get_api_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings & Health"
        ],
        "summary": "Update Api Settings",
        "description": "Update API settings\n\nArgs:\n    payload: Settings to update\n\nReturns:\n    Updated settings",
        "operationId": "update_api_settings_api_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/solscan-settings": {
      "get": {
        "tags": [
          "Settings & Health"
        ],
        "summary": "Get Solscan Settings",
        "description": "Get current Solscan URL settings from action_wheel_settings.ini\n\nReturns:\n    Solscan settings dictionary",
        "operationId": "get_solscan_settings_api_solscan_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings & Health"
        ],
        "summary": "Update Solscan Settings",
        "description": "Update Solscan URL settings in action_wheel_settings.ini\n\nArgs:\n    payload: Solscan settings to update\n\nReturns:\n    Updated settings",
        "operationId": "update_solscan_settings_api_solscan_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolscanSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/analysis": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "List Analyses",
        "description": "List analysis jobs and completed tokens",
        "operationId": "list_analyses_analysis_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/wallets/{wallet_address}/tags": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "Get Wallet Tags",
        "description": "Get tags for a wallet",
        "operationId": "get_wallet_tags_wallets__wallet_address__tags_get",
        "parameters": [
          {
            "name": "wallet_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletTagsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Add Wallet Tag",
        "description": "Add a tag to a wallet",
        "operationId": "add_wallet_tag_wallets__wallet_address__tags_post",
        "parameters": [
          {
            "name": "wallet_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet Address"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTagRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tags"
        ],
        "summary": "Remove Wallet Tag",
        "description": "Remove a tag from a wallet",
        "operationId": "remove_wallet_tag_wallets__wallet_address__tags_delete",
        "parameters": [
          {
            "name": "wallet_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet Address"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveTagRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/wallets/batch-tags": {
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Get Batch Wallet Tags",
        "description": "Get tags for multiple wallets in one query",
        "operationId": "get_batch_wallet_tags_wallets_batch_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchTagsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tags/{tag}/wallets": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "Get Wallets By Tag",
        "description": "Get all wallets with a specific tag",
        "operationId": "get_wallets_by_tag_tags__tag__wallets_get",
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddTagRequest": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "is_kol": {
            "type": "boolean",
            "title": "Is Kol",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "tag"
        ],
        "title": "AddTagRequest"
      },
...
      "AnalysisJobSummary": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "token_address": {
            "type": "string",
            "title": "Token Address"
          },
          "token_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Name"
          },
          "token_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Symbol"
          },
          "acronym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acronym"
          },
          "wallets_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wallets Found"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "credits_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits Used"
          },
          "results_url": {
            "type": "string",
            "title": "Results Url"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "token_address",
          "results_url"
        ],
        "title": "AnalysisJobSummary",
        "description": "Summary info for analysis job in list view"
      },
      "AnalysisListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/AnalysisJobSummary"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "total",
          "jobs"
        ],
        "title": "AnalysisListResponse",
        "description": "Response for listing analysis jobs"
      },
...
      "BatchTagsRequest": {
        "properties": {
          "addresses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Addresses"
          }
        },
        "type": "object",
        "required": [
          "addresses"
        ],
        "title": "BatchTagsRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Simple message response"
      },
...
      "RemoveTagRequest": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          }
        },
        "type": "object",
        "required": [
          "tag"
        ],
        "title": "RemoveTagRequest"
      },
...
      "UpdateSettingsRequest": {
        "properties": {
          "transactionLimit": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Transactionlimit"
          },
          "minUsdFilter": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minusdfilter"
          },
          "walletCount": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Walletcount"
          },
          "apiRateDelay": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Apiratedelay"
          },
          "maxCreditsPerAnalysis": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxcreditsperanalysis"
          },
          "maxRetries": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxretries"
          }
        },
        "type": "object",
        "title": "UpdateSettingsRequest"
      },
      "UpdateSolscanSettingsRequest": {
        "properties": {
          "activity_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Type"
          },
          "exclude_amount_zero": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Amount Zero"
          },
          "remove_spam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remove Spam"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "token_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Address"
          },
          "page_size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Size"
          }
        },
        "type": "object",
        "title": "UpdateSolscanSettingsRequest",
        "description": "Request model for updating Solscan settings"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WalletTag": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "is_kol": {
            "type": "boolean",
            "title": "Is Kol"
          }
        },
        "type": "object",
        "required": [
          "tag",
          "is_kol"
        ],
        "title": "WalletTag"
      },
      "WalletTagsResponse": {
        "properties": {
          "tags": {
            "items": {
              "$ref": "#/components/schemas/WalletTag"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "tags"
        ],
        "title": "WalletTagsResponse"
      }
    }
  }
}